// libs/modules/global/secrets/enum.ts
export enum PropertyAPIEnvironment {
  PORT = 'PORT_PROPERTY_API',
  URL = 'URL_PROPERTY_API',
}

export enum AuthAPIEnvironment {
  PORT = 'PORT_AUTH_API',
  SECRET_JWT = 'SECRET_JWT',
  JWT_EXPIRES_IN = 'JWT_EXPIRES_IN',
  JWT_REFRESH_EXPIRES_IN = 'JWT_REFRESH_EXPIRES_IN',
  URL = 'URL_AUTH_API',
}

export enum BlockchainAPIEnvironment {
  PORT = 'PORT_BLOCKCHAIN_API',
  URL = 'URL_BLOCKCHAIN_API',
}